isolate the cli into it's own folder
