+++ cli/cmd/generate/generate.go
