+++ cli/cmd/verify/verify.go
